fix(io): Cast XLSB ExternSheet reserve count to size_t for the 32-bit WASM build
The 64-bit std::min result narrowed to size_t only trips
-Wshorten-64-to-32 on the 32-bit WASM target, breaking the WASM build.
The value is already clamped to p.size / kExternSheetEntryBytes, so the
explicit cast is value-safe.

+++ src/io/xlsb/reader.cpp
@@ -479,7 +479,8 @@ Expected<std::vector<XlsbSheetRange>, Error> DecodeExternSheet(const std::vector
     // count (up to 4 billion) cannot force a multi-GB reservation before
     // the per-entry reads run out of bytes and fail.
     constexpr std::size_t kExternSheetEntryBytes = 12U;
-    const std::size_t reservable = std::min<std::uint64_t>(count_or.value(), p.size / kExternSheetEntryBytes);
+    const std::size_t reservable =
+        static_cast<std::size_t>(std::min<std::uint64_t>(count_or.value(), p.size / kExternSheetEntryBytes));
     ranges.reserve(reservable);
     for (std::uint32_t i = 0; i < count_or.value(); ++i) {
       auto sup_book_or = read_u32(p);
